Componentize Role Tags Dropdown in PersonForm
