+++ ui/src/Common/PersonForm/PersonForm.tsx
@@ -17,9 +17,7 @@
 
 import React, {FormEvent, useCallback, useEffect, useState} from 'react';
 import AssignmentClient from 'Services/Api/AssignmentClient';
-import RoleClient from 'Services/Api/RoleClient';
 import PeopleClient from 'Services/Api/PeopleClient';
-import {AxiosResponse} from 'axios';
 import {emptyPerson, isArchived} from 'Services/PersonService';
 import {isActiveProduct, isUnassignedProduct} from 'Services/ProductService';
 import SelectWithNoCreateOption, {
@@ -32,7 +30,7 @@ import {ProductPlaceholderPair} from 'Types/CreateAssignmentRequest';
 import moment from 'moment';
 import FormNotesTextArea from 'Common/FormNotesTextArea/FormNotesTextArea';
 import FormButton from 'Common/FormButton/FormButton';
-import SelectWithCreateOption, {MetadataReactSelectProps} from 'Common/SelectWithCreateOption/SelectWithCreateOption';
+import {MetadataReactSelectProps} from 'Common/SelectWithCreateOption/SelectWithCreateOption';
 import FormTagsField from 'Common/FormTagsField/FormTagsField';
 import PersonTagClient from 'Services/Api/PersonTagClient';
 import {RoleTag, Tag} from 'Types/Tag';
@@ -43,16 +41,15 @@ import {ViewingDateState} from 'State/ViewingDateState';
 import {IsUnassignedDrawerOpenState} from 'State/IsUnassignedDrawerOpenState';
 import {ProductsState} from 'State/ProductsState';
 import {PeopleState} from 'State/PeopleState';
-import useFetchRoles from 'Hooks/useFetchRoles/useFetchRoles';
 import {ModalContentsState} from 'State/ModalContentsState';
 import {CurrentSpaceState} from 'State/CurrentSpaceState';
-import {RoleTagRequest} from 'Types/TagRequest';
 import {Product} from 'Types/Product';
 import {Person} from 'Types/Person';
 import {Assignment} from 'Types/Assignment';
 import CDSIDInput from "./CDSIDInput/CDSIDInput";
 import MarkAsNewCheckbox from "./MarkAsNewCheckbox/MarkAsNewCheckbox";
 import NameInput from "./NameInput/NameInput";
+import RoleTagsDropdown from "./RoleTagsDropdown/RoleTagsDropdown";
 
 import './PersonForm.scss';
 
@@ -72,9 +69,7 @@ function PersonForm({ isEditPersonForm, initiallySelectedProduct, initialPersonN
     const setModalContents = useSetRecoilState(ModalContentsState);
 
     const spaceUuid = currentSpace.uuid || '';
-    const { fetchRoles, roles } = useFetchRoles(spaceUuid);
 
-    const { ROLE_TAGS } = MetadataReactSelectProps;
     const { PERSON_ASSIGN_TO } = MetadataMultiSelectProps;
     const { ARCHIVED_PERSON_ASSIGN_TO } = MetadataMultiSelectProps;
     const [confirmDeleteModal, setConfirmDeleteModal] = useState<JSX.Element | null>(null);
@@ -230,30 +225,6 @@ function PersonForm({ isEditPersonForm, initiallySelectedProduct, initialPersonN
         setPerson((updatingPerson: Person) => ({...updatingPerson, [fieldName]: fieldValue}));
     };
 
-    const updateSpaceRole = (input: string): void => {
-        const roleMatch: RoleTag | undefined = roles.find((role: RoleTag) => role.name === input);
-        updatePersonField('spaceRole', roleMatch);
-    };
-
-    const createOption = (role: RoleTag): Option => {
-        return ({
-            label: role.name,
-            value: role.name,
-            color: role.color?.color,
-        });
-    };
-
-    const handleCreateRole = (inputValue: string): void => {
-        setIsLoading(true);
-        const roleAddRequest: RoleTagRequest = {name: inputValue};
-        RoleClient.add(roleAddRequest, currentSpace).then((response: AxiosResponse) => {
-            const newRole: RoleTag = response.data;
-            fetchRoles();
-            updatePersonField('spaceRole', newRole);
-            setIsLoading(false);
-        });
-    };
-
     const displayRemovePersonModal = (): void => {
         const propsForDeleteConfirmationModal: ConfirmationModalProps = {
             submit: removePerson,
@@ -307,14 +278,11 @@ function PersonForm({ isEditPersonForm, initiallySelectedProduct, initialPersonN
                         updatePersonField('customField1', event.target.value);
                     }}
                 />
-                <SelectWithCreateOption
-                    metadata={ROLE_TAGS}
-                    useColorBadge
-                    value={person.spaceRole && person.spaceRole.name !== '' ? createOption(person.spaceRole) : undefined}
-                    options={roles.map(role => createOption(role))}
-                    onChange={(e): void => updateSpaceRole(e ? (e as Option).value : '')}
-                    onSave={handleCreateRole}
+                <RoleTagsDropdown
+                    spaceRole={person?.spaceRole}
+                    onChange={(role) => updatePersonField('spaceRole', role)}
                     isLoading={isLoading}
+                    setIsLoading={setIsLoading}
                 />
                 <SelectWithNoCreateOption
                     metadata={isArchived(person, viewingDate) ? ARCHIVED_PERSON_ASSIGN_TO : PERSON_ASSIGN_TO}
@@ -353,7 +321,6 @@ function PersonForm({ isEditPersonForm, initiallySelectedProduct, initialPersonN
                     </FormButton>
                 </div>
             </form>
-
             {isEditPersonForm && (
                 <button className="deleteButtonContainer alignSelfCenter deleteLinkColor"
                     data-testid="deletePersonButton"
